Refactor: move inbox service
