+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Simplify sync by not reusing IMAP client
+- Move sync inbox into isolate
+- 
 
 ## [0.8.178] - 2022-10-21
 ### Changed:
