+++ test/blocs/sync/sync_config_cubit_test.dart
@@ -6,7 +6,7 @@ import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/sync_config_service.dart';
 import 'package:lotti/sync/inbox/inbox_service.dart';
-import 'package:lotti/sync/outbox_service.dart';
+import 'package:lotti/sync/outbox/outbox_service.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../mocks/sync_config_test_mocks.dart';
